refactor(types): update tstyche config
