+++ package.json
@@ -72,8 +72,10 @@
     "neostandard": "^0.13.0",
     "tstyche": "^7.0.0"
   },
-  "tsd": {
-    "directory": "test"
+  "tstyche": {
+    "testFileMatch": [
+      "types/*.tst.ts"
+    ]
   },
   "publishConfig": {
     "access": "public"
